Add files via upload
The ur3 file is similar to the one from our toolbox but I added limits to the class. 
The assignment 2 file consists of two functions right now that help move and drop the coin at a certain point. Right now there is no claw on the arm and doesn't have collision detection.